test(mobile): split 6 oversized test/helper files by concern
- factories.ts (909 LOC) → 6 subfiles under factories/ barrel; update 23
  hook test imports to factories/index.js
- row-transforms.test.ts (2011 LOC) → 5 files by entity group under
  data/__tests__/row-transforms/
- SyncProvider.test.tsx (993 LOC) → SyncProvider-lifecycle + SyncProvider-pipeline;
  vi.mock setup block duplicated in each (hoisting constraint)
- opfs-worker.test.ts (857 LOC) → opfs-worker-dispatch + opfs-worker-panic
- trpc-persister-api.test.ts (1213 LOC) → trpc-persister-api-core +
  trpc-persister-api-comms
- import.hooks.test.tsx (757 LOC) → import.hooks-start-resume +
  import.hooks-query-cancel

125 test files, 1366 tests pass. Closes mobile-jxox.
