+++ apps/mobile/src/hooks/__tests__/use-innerworld-canvas.test.tsx
@@ -4,7 +4,7 @@ import { WasmSodiumAdapter } from "@pluralscape/crypto/wasm";
 import { act, waitFor } from "@testing-library/react";
 import { beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { makeRawCanvas } from "../../__tests__/factories.js";
+import { makeRawCanvas } from "../../__tests__/factories/index.js";
 
 import { renderHookWithProviders, TEST_SYSTEM_ID } from "./helpers/render-hook-with-providers.js";
