+++ apps/mobile/src/hooks/__tests__/use-system-settings.test.tsx
@@ -5,7 +5,7 @@ import { brandId } from "@pluralscape/types";
 import { act, waitFor } from "@testing-library/react";
 import { beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { makeRawNomenclature, makeRawSystemSettings } from "../../__tests__/factories.js";
+import { makeRawNomenclature, makeRawSystemSettings } from "../../__tests__/factories/index.js";
 
 import { renderHookWithProviders, TEST_SYSTEM_ID } from "./helpers/render-hook-with-providers.js";
